requirements.txt
lumibot
yfinance
pandas
requests
python-dotenv
